+++ doc/conf.py
@@ -55,7 +55,7 @@ copyright = u'2017-2020, Ken Kundert'
 # built documents.
 #
 # The full version, including alpha/beta/rc tags.
-release = '2.9.1'
+release = '2.9.2'
 # The short X.Y version.
 version = '.'.join(release.split('.')[0:2])